requirements.txt
scikit-learn
openpyxl
pdfminer.six==20221105
cachetools
bs4
pandas
flask
python-dotenv
click